feat: allow setting the logging level and use slog
This is a bit of a mess, and initialization is a headache since we want
to honor the koanf config as well as the command-line flag, but not all
of the global flag info is available right away.

While this works, I don't think we actually want a global logging
package, it should probably just be a logger in the cmd package and
we can pass it through to other packages if we want to. That will be
cleaner separation of concerns without the internal packages having to
be aware of the global logger setup.

I also want to make the cmdOptions usage consistent. Not sure I like the
pattern, but I'll tinker with it. I just wanted to commit this working
version because it took a TON of trial and error.