refactor!: rename all of the media files
I've iterated on this a few times, and I think these terms will be
the least confusing. Instead of --output-dir, it's not --media-dir
and the subdirectories are "incoming" and "outgoing", which should be
universally understandable. I like that they're both the same number of
letters as well.

The term output directory is still used in the combine subcommand code
to match ffmpeg's input/output terminology. I do still need to audit all
of my uses of file/filename/path/dir/dirname and so on for consistency.